errores corregidos ahora sí

+++ app/Http/Controllers/inversion_venta/vehiculosController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\inversion_venta;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\inversion_venta\vehiculos;
+use App\Models\inversion_venta\repuestos;
+use App\Models\inversion_venta\ventas;
 
 class vehiculosController extends Controller
 {
@@ -92,6 +94,22 @@ class vehiculosController extends Controller
                     'error' => 'No se pudo realizar correctamente con este id ' . $id . ''
                 ], 404);
             }
+
+            $repuesto = repuestos::query()->where('fk_vehiculo', '=', $id);
+            if (is_null($repuesto)) {
+                return response()->json([
+                    'error' => 'No se pudo realizar correctamente con este id ' . $id . ''
+                ], 404);
+            }
+
+            $venta = ventas::query()->where('fk_ve_hiculo', '=', $id)->count();
+            if ($venta > 0) {
+                return response()->json([
+                    'error' => 'No se puede eliminar el vehículo, porque está relacionado con una venta'
+                ], 422);
+            }
+
+            $repuesto->delete();
 
             $vehiculo->delete();
             return response()->json([
